🚚 Move wasm code in rust submodules.
